+++ Gemfile
@@ -9,10 +9,7 @@ gem 'uglifier', '>= 1.3.0'
 gem 'jquery-rails'
 gem 'devise'
 gem 'carrierwave'
-gem 'foundation-rails'
-gem 'foundation-icons-sass-rails'
 gem 'font-awesome-rails'
-gem 'foundation-datetimepicker-rails', '~> 0.2.4'
 gem 'pusher'
 gem 'pg_search'
 gem 'twitter'
