+++ app/assets/stylesheets/_body.scss
@@ -5,3 +5,8 @@ body {
 .main-content {
   width: 75%;
 }
+
+.flashGoalCreated {
+  font-size: 20px;
+  padding-top: 15px; 
+}
